requirements.txt
scipy==1.0.0
numpy==1.14.0+mkl
noise==1.2.2
Pillow==5.1.0
